Inspection points are added and executing
They are still sent as reportstate objects. I do not know if we can guarantee order or what other shenanigans we will encounter.

I have yet to add local variables and the full list of paramaeters to log. Perhaps going back to the list where things can be ticked and unticked would be good.